:memo: move headers into one folder, rename file, reorder pages in dropdown navbar
